detail activity merge code with new db

+++ app/src/main/java/com/example/x_comic/models/Feedback.kt
@@ -4,12 +4,12 @@ class Feedback :java.io.Serializable{
     var id: String = ""
     var uid: String = ""
     var bid: String = ""
-    var rating: Float = 0.0F
+    var rating: Double = 0.0
     var feedback: String = ""
 
     constructor() {}
 
-    constructor(id: String, uid: String, bid: String, rating: Float, feedback: String) {
+    constructor(id: String, uid: String, bid: String, rating: Double, feedback: String) {
         this.id = id
         this.uid = uid
         this.bid = bid
